Fix Restaurant Grid test suite
The test suite failed for two reasons:

1. Grid contains cards, that contain Link, that require a Router context

I wrapped the tested element in a MemoryRouter,
so the route is reset between test runs.

2. The test runner finds multiple elements by the same text.

That happens because there are at least two pages in the HTML:
the visible page and the pre-loaded hidden next page.
So this time, the test expects two instances of the text.